refactor: :recycle: file structure
