+++ src/board.go
@@ -178,130 +178,6 @@ func (b board) colorBySquareIndex(s SquareIndex) Color {
 	return colorBySquare(b.squares[s])
 }
 
-func (b board) squareKnightAttackers(side Color, squareIndex SquareIndex) []SquareIndex {
-	attackers := make([]SquareIndex, 0, MAX_SQUARE_KNIGHT_ATTACKERS)
-
-	enemyKnight := BLACK_KNIGHT
-
-	if side == BLACK {
-		enemyKnight = WHITE_KNIGHT
-	}
-
-	for _, dir := range MOVE_DIRECTIONS {
-		for _, moveDist := range KNIGHT_MOVE_DISTS {
-			originIndex := SquareIndex(moveDist*dir) + squareIndex
-			if b.squares[originIndex] == enemyKnight {
-				attackers = append(attackers, originIndex)
-			}
-		}
-	}
-	return attackers
-}
-
-func (b board) squareDiagonalAttackers(side Color, squareIndex SquareIndex) []SquareIndex {
-	attackers := make([]SquareIndex, 0, MAX_SQUARE_DIAGONAL_ATTACKERS)
-
-	pawnAttackDir := POSITIVE_DIR
-
-	enemyPawn := BLACK_PAWN
-	enemyBishop := BLACK_BISHOP
-	enemyQueen := BLACK_QUEEN
-	enemyKing := BLACK_KING
-
-	if side == BLACK {
-		pawnAttackDir = NEGATIVE_DIR
-
-		enemyPawn = WHITE_PAWN
-		enemyBishop = WHITE_BISHOP
-		enemyQueen = WHITE_QUEEN
-		enemyKing = WHITE_KING
-	}
-
-	for _, dir := range MOVE_DIRECTIONS {
-		for _, moveDist := range DIAGONAL_MOVE_DISTS {
-			for i := 1; i < MAX_MOVE_RANGE; i++ {
-				originIndex := SquareIndex(moveDist*dir*i) + squareIndex
-				originSquare := b.squares[originIndex]
-				if originSquare == EMPTY {
-					continue
-				}
-				switch originSquare {
-				case enemyBishop:
-					fallthrough
-				case enemyQueen:
-					attackers = append(attackers, originIndex)
-					break
-				case enemyPawn:
-					if i == 1 && dir*-1 == pawnAttackDir {
-						attackers = append(attackers, originIndex)
-					}
-					break
-				case enemyKing:
-					if i == 1 {
-						attackers = append(attackers, originIndex)
-					}
-					break
-				}
-				// Reached invalid, friendly, or non-diagonal moving piece
-				break
-			}
-		}
-	}
-	return attackers
-}
-
-func (b board) squareCardinalAttackers(side Color, squareIndex SquareIndex) []SquareIndex {
-	attackers := make([]SquareIndex, 0, MAX_SQUARE_CARDINAL_ATTACKERS)
-
-	enemyRook := BLACK_ROOK
-	enemyQueen := BLACK_QUEEN
-	enemyKing := BLACK_KING
-
-	if side == BLACK {
-		enemyRook = WHITE_ROOK
-		enemyQueen = WHITE_QUEEN
-		enemyKing = WHITE_KING
-	}
-
-	for _, dir := range MOVE_DIRECTIONS {
-		for _, moveDist := range CARDINAL_MOVE_DISTS {
-			for i := 1; i < MAX_MOVE_RANGE; i++ {
-				originIndex := SquareIndex(moveDist*dir*i) + squareIndex
-				originSquare := b.squares[originIndex]
-				if originSquare == EMPTY {
-					continue
-				}
-				switch originSquare {
-				case enemyRook:
-					fallthrough
-				case enemyQueen:
-					attackers = append(attackers, originIndex)
-					break
-				case enemyKing:
-					if i == 1 {
-						attackers = append(attackers, originIndex)
-					}
-					break
-				}
-				// Reached invalid, friendly, or non-cardinal moving piece
-				break
-			}
-		}
-	}
-	return attackers
-}
-
-func (b board) squareAttackers(side Color, squareIndex SquareIndex) []SquareIndex {
-	attackers := make([]SquareIndex, 0, MAX_SQUARE_ATTACKERS)
-	knightAttackers := b.squareKnightAttackers(side, squareIndex)
-	attackers = append(attackers, knightAttackers...)
-	diagonalAttackers := b.squareDiagonalAttackers(side, squareIndex)
-	attackers = append(attackers, diagonalAttackers...)
-	cardinalAttackers := b.squareCardinalAttackers(side, squareIndex)
-	attackers = append(attackers, cardinalAttackers...)
-	return attackers
-}
-
 func (b board) toString() string {
 	s := ""
 	sep := "\n_________________________\n"
@@ -366,136 +242,4 @@ func (b board) toString() string {
 	s += sep
 	s += " A  B  C  D  E  F  G  H"
 	return s
-}
-
-func colorBySquare(s Square) Color {
-	c := COLOR_NONE
-	switch s {
-	case WHITE_PAWN:
-		fallthrough
-	case WHITE_KNIGHT:
-		fallthrough
-	case WHITE_BISHOP:
-		fallthrough
-	case WHITE_ROOK:
-		fallthrough
-	case WHITE_QUEEN:
-		fallthrough
-	case WHITE_KING:
-		return WHITE
-	case BLACK_PAWN:
-		fallthrough
-	case BLACK_KNIGHT:
-		fallthrough
-	case BLACK_BISHOP:
-		fallthrough
-	case BLACK_ROOK:
-		fallthrough
-	case BLACK_QUEEN:
-		fallthrough
-	case BLACK_KING:
-		return BLACK
-	}
-	return c
-}
-
-func squareIndexByCoord(s SquareCoord) (SquareIndex, error) {
-	var f File
-	var r Rank
-	coordParts := []rune(s)
-	if len(coordParts) != 2 {
-		return 0, fmt.Errorf("Invalid chess notation: %v", s)
-	}
-	switch strings.ToLower(string(coordParts[0])) {
-	case "a":
-		f = FILE_A
-	case "b":
-		f = FILE_B
-	case "c":
-		f = FILE_C
-	case "d":
-		f = FILE_D
-	case "e":
-		f = FILE_E
-	case "f":
-		f = FILE_F
-	case "g":
-		f = FILE_G
-	case "h":
-		f = FILE_H
-	default:
-		return 0, fmt.Errorf("Invalid chess file: %v", string(coordParts[0]))
-	}
-	switch string(coordParts[1]) {
-	case "1":
-		r = RANK_ONE
-	case "2":
-		r = RANK_TWO
-	case "3":
-		r = RANK_THREE
-	case "4":
-		r = RANK_FOUR
-	case "5":
-		r = RANK_FIVE
-	case "6":
-		r = RANK_SIX
-	case "7":
-		r = RANK_SEVEN
-	case "8":
-		r = RANK_EIGHT
-	default:
-		return 0, fmt.Errorf("Invalid chess rank: %v", string(coordParts[1]))
-
-	}
-	return squareIndexByFileRank(f, r), nil
-}
-
-func squareIndexByFileRank(f File, r Rank) SquareIndex {
-	return (21 + SquareIndex(f)) + SquareIndex(r)*10
-}
-
-func rankBySquareIndex(s SquareIndex) Rank {
-	r := RANK_NONE
-	switch (s / 10) % 10 {
-	case 2:
-		return RANK_EIGHT
-	case 3:
-		return RANK_SEVEN
-	case 4:
-		return RANK_SIX
-	case 5:
-		return RANK_FIVE
-	case 6:
-		return RANK_FOUR
-	case 7:
-		return RANK_THREE
-	case 8:
-		return RANK_TWO
-	case 9:
-		return RANK_ONE
-	}
-	return r
-}
-
-func fileBySquareIndex(s SquareIndex) File {
-	f := FILE_NONE
-	switch s % 10 {
-	case 1:
-		return FILE_A
-	case 2:
-		return FILE_B
-	case 3:
-		return FILE_C
-	case 4:
-		return FILE_D
-	case 5:
-		return FILE_E
-	case 6:
-		return FILE_F
-	case 7:
-		return FILE_G
-	case 8:
-		return FILE_H
-	}
-	return f
 }
